week-1-api-security/day-03-openapi-fuzz-tester/pyproject.toml
[tool.poetry]
name = "day-03-openapi-fuzz-tester"
version = "1.0.0"
description = "OpenAPI schema-driven fuzz tester with 7 mutator classes and response anomaly detection"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python   = "^3.11"
httpx    = "^0.27"
pyyaml   = "^6.0"
fastapi  = "^0.110"
uvicorn  = {extras = ["standard"], version = "^0.27"}
pydantic = "^2.6"

[tool.poetry.dev-dependencies]
pytest      = "^8.1"
pytest-cov  = "^5.0"
pytest-mock = "^3.14"

[tool.pytest.ini_options]
testpaths  = ["fuzzer/tests"]
log_cli    = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["fuzzer"]
omit   = ["*/tests/*", "*/sample_target*"]

[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
